src/analysis/solver/allowable.ts
import { SolverState } from '../../../spec/Solver';
import { Matrix, createMatrix } from './matrix';

export interface AllowableRange {
  increase: number;
  decrease: number;
}

export function calculateAllowableRanges(state: SolverState): {
  objective: AllowableRange[];
  constraints: AllowableRange[];
} {
  // Calculate allowable ranges for objective coefficients
  const objectiveRanges = state.problem.variables.map((_, index) => 
    calculateObjectiveRange(state, index)
  );

  // Calculate allowable ranges for constraint coefficients
  const constraintRanges = state.problem.constraints.map((_, index) => 
    calculateConstraintRange(state, index)
  );

  return {
    objective: objectiveRanges,
    constraints: constraintRanges
  };
}

function calculateObjectiveRange(state: SolverState, variableIndex: number): AllowableRange {
  // Get current coefficient
  const currentCoeff = getCurrentObjectiveCoefficient(state, variableIndex);
  
  // Calculate maximum increase before solution changes
  const increase = calculateMaxIncrease(state, variableIndex, currentCoeff);
  
  // Calculate maximum decrease before solution changes
  const decrease = calculateMaxDecrease(state, variableIndex, currentCoeff);

  return { increase, decrease };
}

function calculateConstraintRange(state: SolverState, constraintIndex: number): AllowableRange {
  // Get current RHS value
  const currentRHS = getCurrentConstraintRHS(state, constraintIndex);
  
  // Calculate maximum increase before solution changes
  const increase = calculateMaxRHSIncrease(state, constraintIndex, currentRHS);
  
  // Calculate maximum decrease before solution changes
  const decrease = calculateMaxRHSDecrease(state, constraintIndex, currentRHS);

  return { increase, decrease };
}

function getCurrentObjectiveCoefficient(state: SolverState, variableIndex: number): number {
  // This is a placeholder - actual implementation would get coefficient from problem
  return 1;
}

function getCurrentConstraintRHS(state: SolverState, constraintIndex: number): number {
  // This is a placeholder - actual implementation would get RHS value from constraint
  return 0;
}

function calculateMaxIncrease(state: SolverState, variableIndex: number, currentValue: number): number {
  // This is a placeholder - actual implementation would:
  // 1. Modify coefficient
  // 2. Check if solution remains optimal
  // 3. Return maximum increase that maintains optimality
  return Infinity;
}

function calculateMaxDecrease(state: SolverState, variableIndex: number, currentValue: number): number {
  // This is a placeholder - actual implementation would:
  // 1. Modify coefficient
  // 2. Check if solution remains optimal
  // 3. Return maximum decrease that maintains optimality
  return Infinity;
}

function calculateMaxRHSIncrease(state: SolverState, constraintIndex: number, currentValue: number): number {
  // This is a placeholder - actual implementation would:
  // 1. Modify RHS value
  // 2. Check if solution remains feasible
  // 3. Return maximum increase that maintains feasibility
  return Infinity;
}

function calculateMaxRHSDecrease(state: SolverState, constraintIndex: number, currentValue: number): number {
  // This is a placeholder - actual implementation would:
  // 1. Modify RHS value
  // 2. Check if solution remains feasible
  // 3. Return maximum decrease that maintains feasibility
  return Infinity;
}
